@@ -6,28 +6,29 @@
     "license": "MIT",
     "authors": [],
     "require": {
-        "php": "^7.2.5",
+        "php": "^7.3",
         "ext-json": "*",
-        "aws/aws-sdk-php-laravel": "^3.4",
-        "doctrine/dbal": "^2.10",
-        "illuminate/support": "^7.0",
-        "kriswallsmith/buzz": "^1.1",
-        "laravel/ui": "^2.0",
-        "mailgun/mailgun-php": "^3.0",
-        "mailjet/mailjet-apiv3-php": "^1.4",
-        "nyholm/psr7": "^1.2",
-        "rap2hpoutre/fast-excel": "^1.7",
-        "sendgrid/sendgrid": "^7.3",
-        "wildbit/postmark-php": "^2.8"
+        "aws/aws-sdk-php-laravel": "^3.6",
+        "doctrine/dbal": "^3.0",
+        "illuminate/support": "^8.0",
+        "kriswallsmith/buzz": "^1.2",
+        "mailgun/mailgun-php": "^3.2",
+        "mailjet/mailjet-apiv3-php": "^1.5",
+        "nyholm/psr7": "^1.3",
+        "rap2hpoutre/fast-excel": "^2.3",
+        "sendgrid/sendgrid": "^7.9",
+        "wildbit/postmark-php": "^4.0"
     },
     "require-dev": {
-        "orchestra/testbench": "^5.0",
-        "phpunit/phpunit": "^8.0",
+        "orchestra/testbench": "^6.0",
+        "phpunit/phpunit": "^9.0",
         "roave/security-advisories": "dev-master"
     },
     "autoload": {
         "psr-4": {
-            "Sendportal\\Base\\": "src"
+            "Sendportal\\Base\\": "src",
+            "Database\\Factories\\": "database/factories/",
+            "Database\\Seeders\\": "database/seeders/"
         }
     },
     "autoload-dev": {
