@@ -1,13 +1,50 @@
-<?php namespace Sendportal\Base\Models;
+<?php
 
+declare(strict_types=1);
+
+namespace Sendportal\Base\Models;
+
+use Carbon\Carbon;
+use Database\Factories\SubscriberFactory;
+use Illuminate\Database\Eloquent\Collection as EloquentCollection;
+use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Relations\BelongsToMany;
 use Illuminate\Database\Eloquent\Relations\HasMany;
 use Ramsey\Uuid\Uuid;
 
+/**
+ * @property int $id
+ * @property int $workspace_id
+ * @property string $hash
+ * @property string $email
+ * @property string|null $first_name
+ * @property string|null $last_name
+ * @property array|null $meta
+ * @property Carbon|null $unsubscribed_at
+ * @property int|null $unsubscribed_event_id
+ * @property Carbon|null $created_at
+ * @property Carbon|null $updated_at
+ *
+ * @property EloquentCollection $segments
+ * @property EloquentCollection $messages
+ * 
+ * @property-read string $full_name
+ *
+ * @method static SubscriberFactory factory
+ */
 class Subscriber extends BaseModel
 {
+    use HasFactory;
+
+    protected static function newFactory()
+    {
+        return SubscriberFactory::new();
+    }
+
+    /** @var string */
     protected $table = 'sendportal_subscribers';
 
+    /** @var string[] */
     protected $fillable = [
         'hash',
         'email',
@@ -18,23 +55,10 @@ class Subscriber extends BaseModel
         'unsubscribe_event_id'
     ];
 
-    protected $dates = [
-        'unsubscribed_at',
+    /** @var string[] */
+    protected $casts = [
+        'unsubscribed_at' => 'datetime',
     ];
-
-    protected static function boot()
-    {
-        parent::boot();
-
-        static::creating(function ($model) {
-            $model->hash = Uuid::uuid4()->toString();
-        });
-
-        static::deleting(function (self $subscriber) {
-            $subscriber->segments()->detach();
-            $subscriber->messages()->delete();
-        });
-    }
 
     public function segments(): BelongsToMany
     {
@@ -50,5 +74,25 @@ class Subscriber extends BaseModel
     public function getFullNameAttribute(): string
     {
         return "{$this->first_name} {$this->last_name}";
+    }
+
+    protected static function boot()
+    {
+        parent::boot();
+
+        static::creating(
+            function ($model)
+            {
+                $model->hash = Uuid::uuid4()->toString();
+            }
+        );
+
+        static::deleting(
+            function (self $subscriber)
+            {
+                $subscriber->segments()->detach();
+                $subscriber->messages()->delete();
+            }
+        );
     }
 }
