@@ -20,7 +20,7 @@ class CampaignCancellationControllerTest extends TestCase
     /** @test */
     public function the_confirm_cancel_endpoint_returns_the_confirm_cancel_view()
     {
-        $campaign = factory(Campaign::class)->state('queued')->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        $campaign = Campaign::factory()->queued()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
         $response = $this->get(route('sendportal.campaigns.confirm-cancel', ['id' => $campaign->id]));
 
@@ -30,7 +30,7 @@ class CampaignCancellationControllerTest extends TestCase
     /** @test */
     public function the_cancel_endpoint_cancels_a_queued_campaign()
     {
-        $campaign = factory(Campaign::class)->state('queued')->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        $campaign = Campaign::factory()->queued()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
         $response = $this->post(route('sendportal.campaigns.cancel', ['id' => $campaign->id]));
 
@@ -42,7 +42,7 @@ class CampaignCancellationControllerTest extends TestCase
     /** @test */
     public function the_cancel_endpoint_cancels_a_sending_campaign()
     {
-        $campaign = factory(Campaign::class)->state('sending')->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        $campaign = Campaign::factory()->sending()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
         $response = $this->post(route('sendportal.campaigns.cancel', ['id' => $campaign->id]));
 
@@ -53,7 +53,7 @@ class CampaignCancellationControllerTest extends TestCase
     /** @test */
     public function the_cancel_endpoint_does_not_allow_a_draft_campaign_to_be_cancelled()
     {
-        $campaign = factory(Campaign::class)->state('draft')->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        $campaign = Campaign::factory()->draft()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
         $response = $this->post(route('sendportal.campaigns.cancel', ['id' => $campaign->id]));
 
@@ -65,7 +65,7 @@ class CampaignCancellationControllerTest extends TestCase
     /** @test */
     public function the_cancel_endpoint_does_not_allow_a_sent_campaign_to_be_cancelled()
     {
-        $campaign = factory(Campaign::class)->state('sent')->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        $campaign = Campaign::factory()->sent()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
         $response = $this->post(route('sendportal.campaigns.cancel', ['id' => $campaign->id]));
 
@@ -77,7 +77,7 @@ class CampaignCancellationControllerTest extends TestCase
     /** @test */
     public function the_cancel_endpoint_does_not_allow_a_cancelled_campaign_to_be_cancelled()
     {
-        $campaign = factory(Campaign::class)->state('cancelled')->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        $campaign = Campaign::factory()->cancelled()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
         $response = $this->post(route('sendportal.campaigns.cancel', ['id' => $campaign->id]));
 
@@ -89,16 +89,16 @@ class CampaignCancellationControllerTest extends TestCase
     /** @test */
     public function when_a_sending_send_to_all_campaign_is_cancelled_the_user_is_told_how_many_messages_were_dispatched()
     {
-        $campaign = factory(Campaign::class)->state('sending')->create([
+        $campaign = Campaign::factory()->sending()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
             'save_as_draft' => 0,
             'send_to_all' => 1,
         ]);
 
         // Dispatched
-        factory(Message::class)->create([
+        Message::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
-            'subscriber_id' => factory(Subscriber::class)->create([
+            'subscriber_id' => Subscriber::factory()->create([
                 'workspace_id' => Sendportal::currentWorkspaceId(),
             ])->id,
             'source_id' => $campaign->id,
@@ -106,9 +106,9 @@ class CampaignCancellationControllerTest extends TestCase
         ]);
 
         // Not Sent
-        factory(Message::class)->create([
+        Message::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
-            'subscriber_id' => factory(Subscriber::class)->create([
+            'subscriber_id' => Subscriber::factory()->create([
                 'workspace_id' => Sendportal::currentWorkspaceId(),
             ])->id,
             'source_id' => $campaign->id,
@@ -123,10 +123,10 @@ class CampaignCancellationControllerTest extends TestCase
     /** @test */
     public function when_a_sending_not_send_to_all_campaign_is_cancelled_the_user_is_told_how_many_messages_were_dispatched()
     {
-        $segment = factory(Segment::class)->create([
+        $segment = Segment::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
-        $campaign = factory(Campaign::class)->state('sending')->create([
+        $campaign = Campaign::factory()->sending()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
             'save_as_draft' => 0,
             'send_to_all' => 0,
@@ -134,11 +134,11 @@ class CampaignCancellationControllerTest extends TestCase
         $campaign->segments()->attach($segment->id);
 
         // Dispatched
-        $subscriber = factory(Subscriber::class)->create([
+        $subscriber = Subscriber::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
         $subscriber->segments()->attach($segment->id);
-        factory(Message::class)->create([
+        Message::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
             'subscriber_id' => $subscriber->id,
             'source_id' => $campaign->id,
@@ -146,11 +146,11 @@ class CampaignCancellationControllerTest extends TestCase
         ]);
 
         // Not Sent
-        $otherSubscriber = factory(Subscriber::class)->create([
+        $otherSubscriber = Subscriber::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
         $otherSubscriber->segments()->attach($segment->id);
-        factory(Message::class)->create([
+        Message::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
             'subscriber_id' => $otherSubscriber->id,
             'source_id' => $campaign->id,
@@ -165,22 +165,22 @@ class CampaignCancellationControllerTest extends TestCase
     /** @test */
     public function campaigns_that_save_as_draft_cannot_be_cancelled_until_every_draft_message_has_been_created()
     {
-        $campaign = factory(Campaign::class)->state('sending')->create([
+        $campaign = Campaign::factory()->sending()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
             'send_to_all' => 0,
             'save_as_draft' => 1,
         ]);
-        $segment = factory(Segment::class)->create([
+        $segment = Segment::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
         $campaign->segments()->attach($segment->id);
-        $subscribers = factory(Subscriber::class, 5)->create([
+        $subscribers = Subscriber::factory()->count(5)->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
         $segment->subscribers()->attach($subscribers->pluck('id'));
 
         // Message Drafts
-        factory(Message::class, 3)->state('pending')->create([
+        Message::factory()->count(3)->pending()->create([
             'source_id' => $campaign->id,
         ]);
 
@@ -195,22 +195,22 @@ class CampaignCancellationControllerTest extends TestCase
     /** @test */
     public function campaigns_that_save_as_draft_can_be_cancelled_if_every_draft_message_has_been_created()
     {
-        $campaign = factory(Campaign::class)->state('sending')->create([
+        $campaign = Campaign::factory()->sending()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
             'send_to_all' => 0,
             'save_as_draft' => 1,
         ]);
-        $segment = factory(Segment::class)->create([
+        $segment = Segment::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
         $campaign->segments()->attach($segment->id);
-        $subscribers = factory(Subscriber::class, 5)->create([
+        $subscribers = Subscriber::factory()->count(5)->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
         $segment->subscribers()->attach($subscribers->pluck('id'));
 
         // Message Drafts
-        factory(Message::class, $subscribers->count())->state('pending')->create([
+        Message::factory()->count($subscribers->count())->pending()->create([
             'source_id' => $campaign->id,
         ]);
 
