@@ -20,7 +20,8 @@ class CampaignControllerTest extends TestCase
     /** @test */
     public function the_index_of_campaigns_is_accessible_to_authenticated_users()
     {
-        factory(Campaign::class, 3)->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        // given
+        Campaign::factory()->count(3)->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
         // when
         $response = $this->get(route('sendportal.campaigns.index'));
@@ -57,7 +58,8 @@ class CampaignControllerTest extends TestCase
     /** @test */
     public function the_preview_view_is_accessible_by_authenticated_users()
     {
-        $campaign = factory(Campaign::class)->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        // given
+        $campaign = Campaign::factory()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
         // when
         $response = $this->get(route('sendportal.campaigns.preview', $campaign->id));
@@ -69,7 +71,8 @@ class CampaignControllerTest extends TestCase
     /** @test */
     public function the_edit_view_is_accessible_by_authenticated_users()
     {
-        $campaign = factory(Campaign::class)->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        // given
+        $campaign = Campaign::factory()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
         // when
         $response = $this->get(route('sendportal.campaigns.edit', $campaign->id));
@@ -81,7 +84,8 @@ class CampaignControllerTest extends TestCase
     /** @test */
     public function a_campaign_is_updateable_by_authenticated_users()
     {
-        $campaign = factory(Campaign::class)->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        // given
+        $campaign = Campaign::factory()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
         $campaignUpdateData = [
             'name' => $this->faker->word,
@@ -108,6 +112,7 @@ class CampaignControllerTest extends TestCase
     /** @test */
     public function campaigns_can_be_set_to_not_track_opens()
     {
+        // given
         $campaignStoreData = $this->generateCampaignStoreData();
 
         // when
@@ -124,6 +129,7 @@ class CampaignControllerTest extends TestCase
     /** @test */
     public function campaigns_can_be_set_to_track_opens()
     {
+        // given
         $campaignStoreData = $this->generateCampaignStoreData() + ['is_open_tracking' => true];
 
         // when
@@ -140,6 +146,7 @@ class CampaignControllerTest extends TestCase
     /** @test */
     public function campaigns_can_be_set_to_not_track_clicks()
     {
+        // given
         $campaignStoreData = $this->generateCampaignStoreData();
 
         // when
@@ -156,6 +163,7 @@ class CampaignControllerTest extends TestCase
     /** @test */
     public function campaigns_can_be_set_to_track_clicks()
     {
+        // given
         $campaignStoreData = $this->generateCampaignStoreData() + ['is_click_tracking' => true];
 
         // when
@@ -172,32 +180,38 @@ class CampaignControllerTest extends TestCase
     /** @test */
     public function campaign_content_is_required_if_no_template_is_selected()
     {
+        // given
         $campaignStoreData = $this->generateCampaignStoreData([
             'template_id' => null,
             'content' => null,
         ]);
 
+        // when
         $response = $this->post(route('sendportal.campaigns.store'), $campaignStoreData);
 
+        // then
         $response->assertSessionHasErrors('content');
     }
 
     /** @test */
     public function campaign_content_is_not_required_if_a_template_is_selected()
     {
+        // given
         $campaignStoreData = $this->generateCampaignStoreData([
             'content' => null,
         ]);
 
+        // when
         $response = $this->post(route('sendportal.campaigns.store'), $campaignStoreData);
 
+        // then
         $response->assertSessionHasNoErrors();
     }
 
     private function generateCampaignStoreData(array $overrides = []): array
     {
-        $emailService = factory(EmailService::class)->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
-        $template = factory(Template::class)->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        $emailService = EmailService::factory()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        $template = Template::factory()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
         return array_merge([
             'name' => $this->faker->word,
