@@ -17,11 +17,11 @@ class MessagesControllerTest extends TestCase
     /** @test */
     public function the_index_of_sent_messages_is_accessible_to_an_authenticated_user()
     {
-        factory(Message::class, 3)->create(['workspace_id' => Sendportal::currentWorkspaceId(), 'sent_at' => now()]);
+        // given
+        Message::factory()->count(3)->create(['workspace_id' => Sendportal::currentWorkspaceId(), 'sent_at' => now()]);
 
         // when
-        $response = $this
-            ->get(route('sendportal.messages.index'));
+        $response = $this->get(route('sendportal.messages.index'));
 
         // then
         $response->assertOk();
@@ -30,11 +30,11 @@ class MessagesControllerTest extends TestCase
     /** @test */
     public function the_index_of_draft_messages_is_accessible_to_an_authenticated_user()
     {
-        factory(Message::class, 3)->create(['workspace_id' => Sendportal::currentWorkspaceId(), 'sent_at' => null]);
+        // given
+        Message::factory()->count(3)->create(['workspace_id' => Sendportal::currentWorkspaceId(), 'sent_at' => null]);
 
         // when
-        $response = $this
-            ->get(route('sendportal.messages.draft'));
+        $response = $this->get(route('sendportal.messages.draft'));
 
         // then
         $response->assertOk();
@@ -43,18 +43,17 @@ class MessagesControllerTest extends TestCase
     /** @test */
     public function a_draft_message_can_be_viewed_by_an_authenticated_user()
     {
-        $campaign = factory(Campaign::class)->state('withContent')->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        // given
+        $campaign = Campaign::factory()->withContent()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
-        /** @var Message $message */
-        $message = factory(Message::class)->create([
+        $message = Message::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
             'source_id' => $campaign->id,
             'sent_at' => null
         ]);
 
         // when
-        $response = $this
-            ->get(route('sendportal.messages.show', $message->id));
+        $response = $this->get(route('sendportal.messages.show', $message->id));
 
         // then
         $response->assertOk();
@@ -63,38 +62,42 @@ class MessagesControllerTest extends TestCase
     /** @test */
     public function a_draft_message_can_be_deleted()
     {
-        $campaign = factory(Campaign::class)->state('withContent')->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        // given
+        $campaign = Campaign::factory()->withContent()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
-        /** @var Message $message */
-        $message = factory(Message::class)->create([
+        $message = Message::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
             'source_id' => $campaign->id,
             'sent_at' => null
         ]);
 
+        // when
         $this->delete(route('sendportal.messages.delete', $message->id))
             ->assertRedirect(route('sendportal.messages.draft'));
 
+        // then
         $this->assertDatabaseMissing('sendportal_messages', ['id' => $message->id]);
     }
 
     /** @test */
     public function a_sent_message_cannot_be_deleted()
     {
-        $campaign = factory(Campaign::class)->state('withContent')->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        // given
+        $campaign = Campaign::factory()->withContent()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
-        /** @var Message $message */
-        $message = factory(Message::class)->create([
+        $message = Message::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
             'source_id' => $campaign->id,
             'sent_at' => now()
         ]);
 
+        // when
         $this
             ->from(route('sendportal.messages.draft'))
             ->delete(route('sendportal.messages.delete', $message->id))
             ->assertRedirect(route('sendportal.messages.draft'));
 
+        // then
         $this->assertDatabaseHas('sendportal_messages', ['id' => $message->id]);
     }
 
@@ -104,30 +107,31 @@ class MessagesControllerTest extends TestCase
      */
     public function a_message_can_be_sent_when_other_messages_have_been_sent()
     {
+        // given
         $workspaceId = Sendportal::currentWorkspaceId();
 
-        $campaign = factory(Campaign::class)->state('withContent')->create(['workspace_id' => $workspaceId]);
+        $campaign = Campaign::factory()->withContent()->create(['workspace_id' => $workspaceId]);
 
-        // Message already sent
-        factory(Message::class)->create([
+        Message::factory()->create([
             'workspace_id' => $workspaceId,
             'source_id' => $campaign->id,
-            'sent_at' => now()
+            'sent_at' => now(), // Message already sent.
         ]);
 
-        /** @var Message $message */
-        $draftMessage = factory(Message::class)->create([
+        $draftMessage = Message::factory()->create([
             'workspace_id' => $workspaceId,
             'source_id' => $campaign->id,
             'queued_at' => now(),
         ]);
 
+        // when
         $this->post(route('sendportal.messages.send'), ['id' => $draftMessage->id])
             ->assertRedirect(route('sendportal.messages.draft'))
             ->assertSessionHas('success');
 
         $draftMessage->refresh();
 
-        $this->assertNotNull($draftMessage->sent_at);
+        // then
+        self::assertNotNull($draftMessage->sent_at);
     }
 }
