@@ -38,27 +38,35 @@ class SubscribersImportControllerTest extends TestCase
     /** @test */
     public function it_should_not_allow_any_file_other_than_a_csv_file_to_be_uploaded()
     {
+        // given
         $file = UploadedFile::fake()->image('subscribers.jpg');
 
-        $this->post(route('sendportal.subscribers.import.store'), [
-                'file' => $file
-            ])
-            ->assertSessionHasErrors('file');
+        // when
+        $response = $this->post(route('sendportal.subscribers.import.store'), [
+            'file' => $file
+        ]);
+
+        // then
+        $response->assertSessionHasErrors('file');
     }
 
     /** @test */
     public function it_should_store_the_uploaded_file_process_the_subscribers_and_redirect_to_subscribers_index_page_if_there_are_no_errors()
     {
+        // given
         $file = $this->createFakeCsvFile([
             ['', 'test@email.com', 'Test Name', 'Test Surname']
         ]);
 
-        $this
-            ->post(route('sendportal.subscribers.import.store'), [
-                'file' => $file
-            ])
-            ->assertRedirect(route('sendportal.subscribers.index'))
-            ->assertSessionHas('success');
+        // when
+        $response = $this->post(route('sendportal.subscribers.import.store'), [
+            'file' => $file
+        ]);
+
+        // then
+        $response->assertRedirect(route('sendportal.subscribers.index'));
+
+        $response->assertSessionHas('success');
 
         $this->assertDatabaseHas('sendportal_subscribers', [
             'workspace_id' => Sendportal::currentWorkspaceId(),
@@ -71,18 +79,23 @@ class SubscribersImportControllerTest extends TestCase
     /** @test */
     public function it_should_not_import_subscribers_as_the_upload_has_to_be_validated_before_storing_subscribers()
     {
+        // given
         $file = $this->createFakeCsvFile([
             ['', 'wrong_email', 'Foo', 'Bar'],                  // bad
             ['', 'test@email.com', 'Test Name', 'Test Surname'] // good
         ]);
 
-        $this
+        // when
+        $response = $this
             ->from(route('sendportal.subscribers.import'))
             ->post(route('sendportal.subscribers.import.store'), [
                 'file' => $file
-            ])
-            ->assertRedirect(route('sendportal.subscribers.import'))
-            ->assertSessionHas('errors');
+            ]);
+
+        // then
+        $response->assertRedirect(route('sendportal.subscribers.import'));
+
+        $response->assertSessionHas('errors');
 
         $this->assertDatabaseCount('sendportal_subscribers', 0);
     }
@@ -90,7 +103,8 @@ class SubscribersImportControllerTest extends TestCase
     /** @test */
     public function it_should_allow_attaching_subscribers_to_an_existing_segment()
     {
-        $segment = factory(Segment::class)->create([
+        // given
+        $segment = Segment::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
 
@@ -98,13 +112,16 @@ class SubscribersImportControllerTest extends TestCase
             ['', 'test@email.com', 'Test Name', 'Test Surname']
         ]);
 
-        $this
+        // when
+        $response = $this
             ->post(route('sendportal.subscribers.import.store'), [
                 'file' => $file,
                 'segments' => [$segment->id]
-            ])
-            ->assertRedirect(route('sendportal.subscribers.index'))
-            ->assertSessionHas('success');
+            ]);
+
+        // then
+        $response->assertRedirect(route('sendportal.subscribers.index'));
+        $response->assertSessionHas('success');
 
         $this->assertDatabaseCount('sendportal_subscribers', 1);
         $this->assertDatabaseCount('sendportal_segment_subscriber', 1);
@@ -113,7 +130,8 @@ class SubscribersImportControllerTest extends TestCase
     /** @test */
     public function it_should_allow_attaching_subscribers_to_multiple_segment()
     {
-        $segments = factory(Segment::class, 2)->create([
+        // given
+        $segments = Segment::factory()->count(2)->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
 
@@ -121,13 +139,16 @@ class SubscribersImportControllerTest extends TestCase
             ['', 'test@email.com', 'Test Name', 'Test Surname']
         ]);
 
-        $this
+        // when
+        $response = $this
             ->post(route('sendportal.subscribers.import.store'), [
                 'file' => $file,
                 'segments' => $segments->pluck('id')->toArray()
-            ])
-            ->assertRedirect(route('sendportal.subscribers.index'))
-            ->assertSessionHas('success');
+            ]);
+
+        // then
+        $response->assertRedirect(route('sendportal.subscribers.index'));
+        $response->assertSessionHas('success');
 
         $this->assertDatabaseCount('sendportal_subscribers', 1);
         $this->assertDatabaseCount('sendportal_segment_subscriber', 2);
@@ -136,7 +157,8 @@ class SubscribersImportControllerTest extends TestCase
     /** @test */
     public function it_should_allow_updating_an_existing_subscriber_if_the_id_column_is_filled()
     {
-        $subscriber = factory(Subscriber::class)->create([
+        // given
+        $subscriber = Subscriber::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
 
@@ -144,24 +166,28 @@ class SubscribersImportControllerTest extends TestCase
             [$subscriber->id, 'test@email.com', 'Test Name', 'Test Surname']
         ]);
 
-        $this
+        // when
+        $response = $this
             ->post(route('sendportal.subscribers.import.store'), [
                 'file' => $file
-            ])
-            ->assertRedirect(route('sendportal.subscribers.index'))
-            ->assertSessionHas('success');
+            ]);
 
         $subscriber->refresh();
 
-        $this->assertEquals('test@email.com', $subscriber->email);
-        $this->assertEquals('Test Name', $subscriber->first_name);
-        $this->assertEquals('Test Surname', $subscriber->last_name);
+        // then
+        $response->assertRedirect(route('sendportal.subscribers.index'));
+        $response->assertSessionHas('success');
+
+        self::assertEquals('test@email.com', $subscriber->email);
+        self::assertEquals('Test Name', $subscriber->first_name);
+        self::assertEquals('Test Surname', $subscriber->last_name);
     }
 
     /** @test */
     public function it_should_allow_updating_an_existing_subscriber_by_email_if_it_already_exist()
     {
-        $subscriber = factory(Subscriber::class)->create([
+        // given
+        $subscriber = Subscriber::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
 
@@ -169,23 +195,27 @@ class SubscribersImportControllerTest extends TestCase
             ['', $subscriber->email, 'Test Name', 'Test Surname']
         ]);
 
-        $this
+        // when
+        $response = $this
             ->post(route('sendportal.subscribers.import.store'), [
                 'file' => $file
-            ])
-            ->assertRedirect(route('sendportal.subscribers.index'))
-            ->assertSessionHas('success');
+            ]);
 
         $subscriber->refresh();
 
-        $this->assertEquals('Test Name', $subscriber->first_name);
-        $this->assertEquals('Test Surname', $subscriber->last_name);
+        // then
+        $response->assertRedirect(route('sendportal.subscribers.index'));
+        $response->assertSessionHas('success');
+
+        self::assertEquals('Test Name', $subscriber->first_name);
+        self::assertEquals('Test Surname', $subscriber->last_name);
     }
 
     /** @test */
     public function it_should_notify_the_user_of_how_many_subscribers_have_been_imported_and_how_many_have_been_updated()
     {
-        $subscriber = factory(Subscriber::class)->create([
+        // given
+        $subscriber = Subscriber::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
 
@@ -195,28 +225,32 @@ class SubscribersImportControllerTest extends TestCase
             ['', 'test2@email.com', 'Test Name', 'Test Surname'] // create
         ]);
 
-        $this
+        // when
+        $response = $this
             ->post(route('sendportal.subscribers.import.store'), [
                 'file' => $file
-            ])
-            ->assertRedirect(route('sendportal.subscribers.index'))
-            ->assertSessionHas('success', 'Imported 2 subscriber(s) and updated 1 subscriber(s) out of 3');
+            ]);
+
+        // then
+        $response->assertRedirect(route('sendportal.subscribers.index'));
+        $response->assertSessionHas('success', 'Imported 2 subscriber(s) and updated 1 subscriber(s) out of 3');
     }
 
     /** @test */
     public function it_should_not_remove_existing_segments_from_a_subscriber_when_importing_subscribers_with_a_new_segment()
     {
-        $subscriber = factory(Subscriber::class)->create([
+        // given
+        $subscriber = Subscriber::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
 
         $subscriber->segments()->attach(
-            factory(Segment::class)->create([
+            Segment::factory()->create([
                 'workspace_id' => Sendportal::currentWorkspaceId(),
             ])
         );
 
-        $segment = factory(Segment::class)->create([
+        $segment = Segment::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
 
@@ -224,24 +258,28 @@ class SubscribersImportControllerTest extends TestCase
             ['', $subscriber->email, 'Test Name', 'Test Surname']
         ]);
 
-        $this
+        // when
+        $response = $this
             ->post(route('sendportal.subscribers.import.store'), [
                 'file' => $file,
                 'segments' => [$segment->id]
-            ])
-            ->assertRedirect(route('sendportal.subscribers.index'))
-            ->assertSessionHas('success');
+            ]);
 
         $subscriber->refresh();
+
+        // then
+        $response->assertRedirect(route('sendportal.subscribers.index'));
+        $response->assertSessionHas('success');
 
         $this->assertDatabaseCount('sendportal_segment_subscriber', 2);
     }
 
     protected function createFakeCsvFile(array $rows)
     {
-        $rows = array_map(function ($row) {
-            return implode(',', $row);
-        }, $rows);
+        $rows = array_map(
+            static function ($row) {
+                return implode(',', $row);
+            }, $rows);
 
         array_unshift($rows, implode(',', ['id', 'email', 'first_name', 'last_name']));
 
