@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tests\Feature\Templates;
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
@@ -14,28 +16,28 @@ class TemplatesControllerTest extends TestCase
     use RefreshDatabase,
         WithFaker;
 
-    protected function setUp(): void
-    {
-        parent::setUp(); // TODO: Change the autogenerated stub
-    }
-
     /** @test */
     public function a_logged_in_user_can_see_template_index()
     {
+        // when
         $response = $this->get(route('sendportal.templates.index'));
 
+        // then
         $response->assertOk();
     }
 
     /** @test */
     public function the_index_lists_existing_templates()
     {
-        $template = factory(Template::class)->create([
+        // given
+        $template = Template::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId()
         ]);
 
+        // when
         $response = $this->get(route('sendportal.templates.index'));
 
+        // then
         $response->assertOk();
         $response->assertSee($template->name);
     }
@@ -43,8 +45,10 @@ class TemplatesControllerTest extends TestCase
     /** @test */
     public function a_logged_in_user_can_see_the_create_form()
     {
+        // when
         $response = $this->get(route('sendportal.templates.create'));
 
+        // then
         $response->assertOk();
         $response->assertSee('New Template');
         $response->assertSee('Template Name');
@@ -54,13 +58,16 @@ class TemplatesControllerTest extends TestCase
     /** @test */
     public function a_logged_in_user_can_store_a_new_template()
     {
+        // given
         $data = [
             'name' => $this->faker->name,
             'content' => $this->faker->sentence
         ];
 
+        // when
         $response = $this->post(route('sendportal.templates.store'), $data);
 
+        // then
         $response->assertRedirect(route('sendportal.templates.index'));
 
         $this->assertDatabaseHas('sendportal_templates', [
@@ -73,32 +80,40 @@ class TemplatesControllerTest extends TestCase
     /** @test */
     public function storing_is_validated()
     {
-        $data = [
+        // given
+        $namePostData = [
             'name' => $this->faker->name,
         ];
 
-        $response = $this->post(route('sendportal.templates.store'), $data);
-
-        $response->assertSessionHasErrors('content');
-
-        $data = [
+        $contentPostData = [
             'content' => $this->faker->sentence
         ];
 
-        $response = $this->post(route('sendportal.templates.store'), $data);
+        // when
+        $namePostResponse = $this->post(route('sendportal.templates.store'), $namePostData);
 
-        $response->assertSessionHasErrors('name');
+        // then
+        $namePostResponse->assertSessionHasErrors('content');
+
+        // when
+        $contentPostResponse = $this->post(route('sendportal.templates.store'), $contentPostData);
+
+        // then
+        $contentPostResponse->assertSessionHasErrors('name');
     }
 
     /** @test */
     public function a_logged_in_user_can_see_the_edit_form()
     {
-        $template = factory(Template::class)->create([
+        // given
+        $template = Template::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId()
         ]);
 
+        // when
         $response = $this->get(route('sendportal.templates.edit', $template->id));
 
+        // then
         $response->assertOk();
 
         $response->assertSee($template->name);
@@ -108,7 +123,8 @@ class TemplatesControllerTest extends TestCase
     /** @test */
     public function a_logged_in_user_can_update_a_template()
     {
-        $template = factory(Template::class)->create([
+        // given
+        $template = Template::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId()
         ]);
 
@@ -117,8 +133,10 @@ class TemplatesControllerTest extends TestCase
             'content' => $this->faker->sentence
         ];
 
+        // when
         $response = $this->put(route('sendportal.templates.update', $template->id), $data);
 
+        // then
         $response->assertRedirect(route('sendportal.templates.index'));
 
         $this->assertDatabaseMissing('sendportal_templates', [
@@ -126,44 +144,51 @@ class TemplatesControllerTest extends TestCase
             'name' => $template->name,
             'content' => $template->content
         ]);
+
         $this->assertDatabaseHas('sendportal_templates', $data + ['id' => $template->id, 'workspace_id' => Sendportal::currentWorkspaceId()]);
     }
 
     /** @test */
     public function updates_are_validated()
     {
-        $template = factory(Template::class)->create([
+        // given
+        $template = Template::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId()
         ]);
 
-        $data = [
+        $namePostData = [
             'name' => $this->faker->name,
         ];
 
-        $response = $this->put(route('sendportal.templates.update', $template->id), $data);
-
-        $response->assertSessionHasErrors('content');
-
-        $data = [
+        $contentPostData = [
             'content' => $this->faker->sentence
         ];
 
-        $response = $this->put(route('sendportal.templates.update', $template->id), $data);
+        // when
+        $namePostResponse = $this->put(route('sendportal.templates.update', $template->id), $namePostData);
+        
+        // then
+        $namePostResponse->assertSessionHasErrors('content');
 
-        $response->assertSessionHasErrors('name');
+        // when
+        $contentPostResponse = $this->put(route('sendportal.templates.update', $template->id), $contentPostData);
+
+        // then
+        $contentPostResponse->assertSessionHasErrors('name');
     }
-
-    // Destroy.
 
     /** @test */
     public function a_logged_in_user_can_delete_a_template()
     {
-        $template = factory(Template::class)->create([
+        // given
+        $template = Template::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId()
         ]);
 
+        // when
         $response = $this->delete(route('sendportal.templates.destroy', $template->id));
 
+        // then
         $response->assertRedirect(route('sendportal.templates.index'));
 
         $this->assertDatabaseMissing('sendportal_templates', [
@@ -175,17 +200,20 @@ class TemplatesControllerTest extends TestCase
     /** @test */
     public function a_logged_in_user_cannot_delete_a_template_if_it_is_used()
     {
-        $template = factory(Template::class)->create([
+        // given
+        $template = Template::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId()
         ]);
 
-        $campaign = factory(Campaign::class)->create([
+        Campaign::factory()->create([
             'template_id' => $template->id
         ]);
 
+        // when
         $response = $this->from(route('sendportal.templates.index'))
             ->delete(route('sendportal.templates.destroy', $template->id));
 
+        // then
         $response->assertRedirect(route('sendportal.templates.index'))
             ->assertSessionHasErrors(['template']);
 
@@ -198,16 +226,20 @@ class TemplatesControllerTest extends TestCase
     /** @test */
     public function a_template_name_must_be_unique_for_a_workspace()
     {
-        $template = factory(Template::class)->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
+        // given
+        $template = Template::factory()->create(['workspace_id' => Sendportal::currentWorkspaceId()]);
 
         $request = [
             'name' => $template->name,
         ];
 
+        // when
         $response = $this->post(route('sendportal.templates.store'), $request);
 
+        // then
         $response->assertRedirect()
             ->assertSessionHasErrors('name');
-        $this->assertEquals(1, Template::where('name', $template->name)->count());
+
+        self::assertEquals(1, Template::where('name', $template->name)->count());
     }
 }
