@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tests\Feature\Webhooks;
 
 use Illuminate\Foundation\Testing\RefreshDatabase;
@@ -13,94 +15,97 @@ class SendgridWebhooksTest extends TestCase
     use RefreshDatabase;
     use WithFaker;
 
-    /**
-     * @var string
-     */
+    /** @var string */
     protected $route = 'sendportal.api.webhooks.sendgrid';
 
-    /**
-     * @return void
-     */
-    public function testDelivered()
+    /** @test */
+    function it_accepts_delivered_webhooks()
     {
+        // given
         $message = $this->createMessage();
 
-        $this->assertNull($message->delivered_at);
+        self::assertNull($message->delivered_at);
 
         $webhook = $this->resolveWebhook('delivered', $message->message_id);
 
+        // when
         $this->json('POST', route($this->route), $webhook);
 
-        $this->assertNotNull($message->refresh()->delivered_at);
+        // then
+        self::assertNotNull($message->refresh()->delivered_at);
     }
 
-    /**
-     * @return void
-     */
-    public function testOpen()
+    /** @test */
+    function it_accepts_open_webhooks()
     {
+        // given
         $message = $this->createMessage();
 
-        $this->assertEquals(0, $message->open_count);
-        $this->assertNull($message->opened_at);
+        self::assertEquals(0, $message->open_count);
+        self::assertNull($message->opened_at);
 
         $webhook = $this->resolveWebhook('open', $message->message_id);
 
+        // when
         $this->json('POST', route($this->route), $webhook);
 
-        $this->assertNotNull($message->refresh()->opened_at);
-        $this->assertEquals(1, $message->open_count);
+        // then
+        self::assertNotNull($message->refresh()->opened_at);
+        self::assertEquals(1, $message->open_count);
     }
 
-    /**
-     * @return void
-     */
-    public function testClick()
+    /** @test */
+    function it_accepts_click_webhooks()
     {
+        // given
         $message = $this->createMessage();
 
-        $this->assertEquals(0, $message->click_count);
-        $this->assertNull($message->clicked_at);
+        self::assertEquals(0, $message->click_count);
+        self::assertNull($message->clicked_at);
 
         $url = ['url' => $this->faker->url];
         $webhook = $this->resolveWebhook('click', $message->message_id, $url);
 
+        // when
         $this->json('POST', route($this->route), $webhook);
 
-        $this->assertNotNull($message->refresh()->clicked_at);
-        $this->assertEquals(1, $message->click_count);
+        // then
+        self::assertNotNull($message->refresh()->clicked_at);
+        self::assertEquals(1, $message->click_count);
     }
 
-    /**
-     * @return void
-     */
-    public function testSpamReport()
+    /** @test */
+    function it_accepts_spam_report_webhooks()
     {
+        // given
         $message = $this->createMessage();
 
-        $this->assertNull($message->unsubscribed_at);
+        self::assertNull($message->unsubscribed_at);
 
         $webhook = $this->resolveWebhook('spamreport', $message->message_id);
 
+        // when
         $this->json('POST', route($this->route), $webhook);
 
-        $this->assertNotNull($message->refresh()->unsubscribed_at);
+        // then
+        self::assertNotNull($message->refresh()->unsubscribed_at);
     }
 
-    /**
-     * @return void
-     */
-    public function testDropped()
+    /** @test */
+    function it_accepts_dropped_webhooks()
     {
+        // given
         $message = $this->createMessage();
 
-        $this->assertNull($message->bounced_at);
+        self::assertNull($message->bounced_at);
 
         $webhook = $this->resolveWebhook('dropped', $message->message_id);
 
+        // when
         $this->json('POST', route($this->route), $webhook);
 
-        $this->assertNotNull($message->refresh()->bounced_at);
+        // then
+        self::assertNotNull($message->refresh()->bounced_at);
 
         $this->assertDatabaseHas(
             'sendportal_message_failures',
@@ -111,17 +116,18 @@ class SendgridWebhooksTest extends TestCase
         );
     }
 
-    /**
-     * @return void
-     */
-    public function testDeferred()
+    /** @test */
+    function it_accepts_deferred_webhooks()
     {
+        // given
         $message = $this->createMessage();
 
         $webhook = $this->resolveWebhook('deferred', $message->message_id);
 
+        // when
         $this->json('POST', route($this->route), $webhook);
 
+        // then
         $this->assertDatabaseHas(
             'sendportal_message_failures',
             [
@@ -131,20 +137,21 @@ class SendgridWebhooksTest extends TestCase
         );
     }
 
-    /**
-     * @return void
-     */
-    public function testBounce()
+    /** @test */
+    function it_accepts_bounce_webhooks()
     {
+        // given
         $message = $this->createMessage();
 
-        $this->assertNull($message->bounced_at);
+        self::assertNull($message->bounced_at);
 
         $webhook = $this->resolveWebhook('bounce', $message->message_id);
 
+        // when
         $this->json('POST', route($this->route), $webhook);
 
-        $this->assertNotNull($message->refresh()->bounced_at);
+        // then
+        self::assertNotNull($message->refresh()->bounced_at);
 
         $this->assertDatabaseHas(
             'sendportal_message_failures',
@@ -155,17 +162,18 @@ class SendgridWebhooksTest extends TestCase
         );
     }
 
-    /**
-     * @return void
-     */
-    public function testBlocked()
+    /** @test */
+    function it_accepts_blocked_webhooks()
     {
+        // given
         $message = $this->createMessage();
 
         $webhook = $this->resolveWebhook('blocked', $message->message_id);
 
+        // when
         $this->json('POST', route($this->route), $webhook);
 
+        // then
         $this->assertDatabaseHas(
             'sendportal_message_failures',
             [
@@ -175,25 +183,26 @@ class SendgridWebhooksTest extends TestCase
         );
     }
 
-    /**
-     * @return void
-     */
-    public function testUnsubscribe()
+    /** @test */
+    function it_accepts_unsubscribed_webhooks()
     {
+        // given
         $message = $this->createMessage();
 
-        $this->assertNull($message->unsubscribed_at);
+        self::assertNull($message->unsubscribed_at);
 
         $webhook = $this->resolveWebhook('unsubscribe', $message->message_id);
 
+        // when
         $this->json('POST', route($this->route), $webhook);
 
-        $this->assertNotNull($message->refresh()->unsubscribed_at);
+        // then
+        self::assertNotNull($message->refresh()->unsubscribed_at);
     }
 
     protected function createMessage(): Message
     {
-        return factory(Message::class)->create(
+        return Message::factory()->create(
             [
                 'message_id' => Str::random(),
             ]
@@ -202,10 +211,12 @@ class SendgridWebhooksTest extends TestCase
 
     protected function resolveWebhook(string $type, string $messageId, array $properties = []): array
     {
-        return [[
-            'event' => $type,
-            'sg_message_id' => $messageId,
-            'timestamp' => now()->timestamp,
-        ] + $properties];
+        return [
+            [
+                'event' => $type,
+                'sg_message_id' => $messageId,
+                'timestamp' => now()->timestamp,
+            ] + $properties
+        ];
     }
 }
