@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tests;
 
 use Sendportal\Base\Facades\Sendportal;
@@ -12,29 +14,32 @@ trait SendportalTestSupportTrait
 {
     protected function createEmailService(): EmailService
     {
-        return factory(EmailService::class)->create([
+        return EmailService::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
     }
 
     protected function createCampaign(EmailService $emailService): Campaign
     {
-        return factory(Campaign::class)->states(['withContent', 'sent'])->create([
-            'workspace_id' => Sendportal::currentWorkspaceId(),
-            'email_service_id' => $emailService->id,
-        ]);
+        return Campaign::factory()
+            ->withContent()
+            ->sent()
+            ->create([
+                'workspace_id' => Sendportal::currentWorkspaceId(),
+                'email_service_id' => $emailService->id,
+            ]);
     }
 
     protected function createSegment(): Segment
     {
-        return factory(Segment::class)->create([
+        return Segment::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
     }
 
     protected function createSubscriber(): Subscriber
     {
-        return factory(Subscriber::class)->create([
+        return Subscriber::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
         ]);
     }
