@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tests\Unit\Repositories;
 
 use Carbon\CarbonImmutable;
@@ -18,9 +20,7 @@ class MessageTenantRepositoryTest extends TestCase
 {
     use RefreshDatabase;
 
-    /**
-     * @var MySqlMessageTenantRepository|PostgresMessageTenantRepository
-     */
+    /** @var MySqlMessageTenantRepository|PostgresMessageTenantRepository */
     protected $repository;
 
     public function setUp(): void
@@ -31,34 +31,38 @@ class MessageTenantRepositoryTest extends TestCase
     }
 
     /** @test */
-    public function it_should_not_count_messages_that_have_not_been_opened_yet()
+    function it_should_not_count_messages_that_have_not_been_opened_yet()
     {
-        $campaign = factory(Campaign::class)->create([
+        // given
+        $campaign = Campaign::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId()
         ]);
 
-        factory(Message::class, 2)->create([
+        Message::factory()->count(2)->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
             'source_id' => $campaign->id,
         ]);
 
-        $data = $this->repository->countUniqueOpensPerPeriod(Sendportal::currentWorkspaceId(), get_class($campaign), $campaign->id, CarbonInterval::day()->totalSeconds);
+        // when
+        $data = $this->repository->countUniqueOpensPerPeriod(Sendportal::currentWorkspaceId(), get_class($campaign), $campaign->id, (int)CarbonInterval::day()->totalSeconds);
 
-        $this->assertInstanceOf(Collection::class, $data);
-        $this->assertTrue($data->isEmpty());
+        // then
+        self::assertInstanceOf(Collection::class, $data);
+        self::assertTrue($data->isEmpty());
     }
 
     /** @test */
-    public function it_should_count_messages_that_have_been_opened_grouped_by_day_period()
+    function it_should_count_messages_that_have_been_opened_grouped_by_day_period()
     {
+        // given
         $opened_at = CarbonImmutable::create(2020, 05, 9, 20);
-        $campaign = factory(Campaign::class)->create([
+        $campaign = Campaign::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId()
         ]);
 
         // 20 - 21 - 22 - 23
         foreach (range(0, 3) as $i) {
-            factory(Message::class)->create([
+            Message::factory()->create([
                 'workspace_id' => Sendportal::currentWorkspaceId(),
                 'source_id' => $campaign->id,
                 'opened_at' => $opened_at->addHours($i)
@@ -68,37 +72,40 @@ class MessageTenantRepositoryTest extends TestCase
         // 24
         $next_opened_at = $opened_at->addHours(4);
 
-        factory(Message::class)->create([
+        Message::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
             'source_id' => $campaign->id,
             'opened_at' => $next_opened_at
         ]);
 
-        $data = $this->repository->countUniqueOpensPerPeriod(Sendportal::currentWorkspaceId(), get_class($campaign), $campaign->id, CarbonInterval::day()->totalSeconds);
+        // when
+        $data = $this->repository->countUniqueOpensPerPeriod(Sendportal::currentWorkspaceId(), get_class($campaign), $campaign->id, (int)CarbonInterval::day()->totalSeconds);
 
-        $this->assertEquals(2, $data->count());
+        // then
+        self::assertEquals(2, $data->count());
 
-        $this->assertEquals(4, $data->first()->open_count);
-        $this->assertEquals($opened_at->toDateTimeString(), $data->first()->opened_at);
-        $this->assertEquals($opened_at->startOfDay()->toDateTimeString(), $data->first()->period_start);
+        self::assertEquals(4, $data->first()->open_count);
+        self::assertEquals($opened_at->toDateTimeString(), $data->first()->opened_at);
+        self::assertEquals($opened_at->startOfDay()->toDateTimeString(), $data->first()->period_start);
 
-        $this->assertEquals(1, $data->last()->open_count);
-        $this->assertEquals($next_opened_at->toDateTimeString(), $data->last()->opened_at);
-        $this->assertEquals($next_opened_at->startOfDay()->toDateTimeString(), $data->last()->period_start);
+        self::assertEquals(1, $data->last()->open_count);
+        self::assertEquals($next_opened_at->toDateTimeString(), $data->last()->opened_at);
+        self::assertEquals($next_opened_at->startOfDay()->toDateTimeString(), $data->last()->period_start);
     }
 
     /** @test */
-    public function it_should_count_messages_that_have_been_opened_grouped_by_two_hours_period()
+    function it_should_count_messages_that_have_been_opened_grouped_by_two_hours_period()
     {
+        // given
         $opened_at = CarbonImmutable::create(2020, 05, 9, 20);
 
-        $campaign = factory(Campaign::class)->create([
+        $campaign = Campaign::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId()
         ]);
 
         // 20 - 21 - 22 - 23
         foreach (range(0, 3) as $i) {
-            factory(Message::class)->create([
+            Message::factory()->create([
                 'workspace_id' => Sendportal::currentWorkspaceId(),
                 'source_id' => $campaign->id,
                 'opened_at' => $opened_at->addHours($i)
@@ -108,44 +115,47 @@ class MessageTenantRepositoryTest extends TestCase
         // 24
         $next_opened_at = $opened_at->addHours(4);
 
-        factory(Message::class)->create([
+        Message::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
             'source_id' => $campaign->id,
             'opened_at' => $next_opened_at
         ]);
 
-        $data = $this->repository->countUniqueOpensPerPeriod(Sendportal::currentWorkspaceId(), get_class($campaign), $campaign->id, CarbonInterval::hours(2)->totalSeconds);
+        // when
+        $data = $this->repository->countUniqueOpensPerPeriod(Sendportal::currentWorkspaceId(), get_class($campaign), $campaign->id, (int)CarbonInterval::hours(2)->totalSeconds);
 
-        $this->assertEquals(3, $data->count());
+        // then
+        self::assertEquals(3, $data->count());
 
         // 20
-        $this->assertEquals(2, $data[0]->open_count);
-        $this->assertEquals($opened_at->toDateTimeString(), $data[0]->opened_at);
-        $this->assertEquals($opened_at->toDateTimeString(), $data[0]->period_start);
+        self::assertEquals(2, $data[0]->open_count);
+        self::assertEquals($opened_at->toDateTimeString(), $data[0]->opened_at);
+        self::assertEquals($opened_at->toDateTimeString(), $data[0]->period_start);
 
         // 22
-        $this->assertEquals(2, $data[0]->open_count);
-        $this->assertEquals($opened_at->addHours(2)->toDateTimeString(), $data[1]->opened_at);
-        $this->assertEquals($opened_at->addHours(2)->toDateTimeString(), $data[1]->period_start);
+        self::assertEquals(2, $data[0]->open_count);
+        self::assertEquals($opened_at->addHours(2)->toDateTimeString(), $data[1]->opened_at);
+        self::assertEquals($opened_at->addHours(2)->toDateTimeString(), $data[1]->period_start);
 
         // 24
-        $this->assertEquals(1, $data[2]->open_count);
-        $this->assertEquals($next_opened_at->toDateTimeString(), $data[2]->opened_at);
-        $this->assertEquals($next_opened_at->startOfDay()->toDateTimeString(), $data[2]->period_start);
+        self::assertEquals(1, $data[2]->open_count);
+        self::assertEquals($next_opened_at->toDateTimeString(), $data[2]->opened_at);
+        self::assertEquals($next_opened_at->startOfDay()->toDateTimeString(), $data[2]->period_start);
     }
 
     /** @test */
-    public function it_should_count_messages_that_have_been_opened_grouped_by_hour_period()
+    function it_should_count_messages_that_have_been_opened_grouped_by_hour_period()
     {
+        // given
         $opened_at = CarbonImmutable::create(2020, 05, 9, 20);
 
-        $campaign = factory(Campaign::class)->create([
+        $campaign = Campaign::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId()
         ]);
 
         // 20 - 21 - 22 - 23
         foreach (range(0, 3) as $i) {
-            factory(Message::class)->create([
+            Message::factory()->create([
                 'workspace_id' => Sendportal::currentWorkspaceId(),
                 'source_id' => $campaign->id,
                 'opened_at' => $opened_at->addHours($i)
@@ -155,23 +165,25 @@ class MessageTenantRepositoryTest extends TestCase
         // 24
         $next_opened_at = $opened_at->addHours(4);
 
-        factory(Message::class)->create([
+        Message::factory()->create([
             'workspace_id' => Sendportal::currentWorkspaceId(),
             'source_id' => $campaign->id,
             'opened_at' => $next_opened_at
         ]);
 
-        $data = $this->repository->countUniqueOpensPerPeriod(Sendportal::currentWorkspaceId(), get_class($campaign), $campaign->id, CarbonInterval::hour()->totalSeconds);
+        // when
+        $data = $this->repository->countUniqueOpensPerPeriod(Sendportal::currentWorkspaceId(), get_class($campaign), $campaign->id, (int)CarbonInterval::hour()->totalSeconds);
 
-        $this->assertEquals(5, $data->count());
+        // then
+        self::assertEquals(5, $data->count());
 
         foreach (range(0, 3) as $i) {
-            $this->assertEquals(1, $data[$i]->open_count);
-            $this->assertEquals($opened_at->addHours($i)->toDateTimeString(), $data[$i]->period_start);
+            self::assertEquals(1, $data[$i]->open_count);
+            self::assertEquals($opened_at->addHours($i)->toDateTimeString(), $data[$i]->period_start);
         }
 
-        $this->assertEquals(1, $data->last()->open_count);
-        $this->assertEquals($next_opened_at->toDateTimeString(), $data->last()->opened_at);
-        $this->assertEquals($next_opened_at->startOfDay()->toDateTimeString(), $data->last()->period_start);
+        self::assertEquals(1, $data->last()->open_count);
+        self::assertEquals($next_opened_at->toDateTimeString(), $data->last()->opened_at);
+        self::assertEquals($next_opened_at->startOfDay()->toDateTimeString(), $data->last()->period_start);
     }
 }
