@@ -1,10 +1,10 @@
 <?php
 
+declare(strict_types=1);
+
 namespace Tests;
 
-use Collective\Html\FormFacade;
 use Illuminate\Support\Str;
-use Laravel\Ui\UiServiceProvider;
 use Orchestra\Testbench\TestCase as BaseTestCase;
 use Sendportal\Base\SendportalBaseServiceProvider;
 use Sendportal\Base\Services\Messages\RelayMessage;
@@ -17,10 +17,9 @@ abstract class TestCase extends BaseTestCase
      * @param \Illuminate\Foundation\Application $app
      * @return array
      */
-    protected function getPackageProviders($app)
+    protected function getPackageProviders($app): array
     {
         return [
-            UiServiceProvider::class,
             SendportalBaseServiceProvider::class,
             SendportalTestServiceProvider::class,
         ];
@@ -35,27 +34,15 @@ abstract class TestCase extends BaseTestCase
 
         $this->withoutMix();
         $this->withExceptionHandling();
-        $this->withFactories(__DIR__ . '/../database/factories');
         $this->mockRelayMessageService();
 
         $this->artisan('migrate')->run();
     }
 
     /**
-     * @param \Illuminate\Foundation\Application $app
-     * @return array
-     */
-    protected function getPackageAliases($app)
-    {
-        return [
-            'Form' => FormFacade::class,
-        ];
-    }
-
-    /**
      * @return void
      */
-    protected function mockRelayMessageService()
+    protected function mockRelayMessageService(): void
     {
         $service = $this->getMockBuilder(RelayMessage::class)
             ->disableOriginalConstructor()
